pyproject.toml
[project]
name = "Bot7685"
version = "0.1.0"
description = "Bot7685 by wyf7685"
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
  # nonebot2
  "nonebot2[fastapi,httpx,websockets,aiohttp]>=2.4.3",
  # adapters
  "nonebot-adapter-discord>=0.1.8",
  "nonebot-adapter-feishu>=2.6.2",
  "nonebot-adapter-milky>=0.4.3",
  "nonebot-adapter-onebot>=2.4.6",
  "nonebot-adapter-qq>=1.6.5",
  "nonebot-adapter-satori>=0.13.7",
  "nonebot-adapter-telegram>=0.1.0b20",
  # plugins
  "nonebot-plugin-akinator>=1.0.3",
  "nonebot-plugin-alconna>=0.59.4",
  "nonebot-plugin-apscheduler>=0.5.0",
  "nonebot-plugin-chatrecorder>=0.7.0",
  "nonebot-plugin-htmlrender==0.6.7",
  "nonebot-plugin-localstore>=0.7.4",
  "nonebot-plugin-orm[asyncpg]>=0.8.2",
  "nonebot-plugin-picstatus>=2.1.4",
  "nonebot-plugin-skland>=0.5.4",
  "nonebot-plugin-uninfo>=0.9.0",
  "nonebot-plugin-user>=0.5.3",
  "nonebot-plugin-waiter>=0.8.1",
  "nonebot-plugin-wakatime>=0.3.0",
  "nonebot-plugin-wordcloud>=0.9.0",
  # plugin deps
  "aiocache[redis]>=0.12.3",
  "anyio>=4.9.0",
  "async-lru>=2.0.5",
  "cloudscraper>=1.2.71",
  "cos-python-sdk-v5>=1.9.36",
  "gunicorn>=23.0.0",
  "httpx>=0.28.1",
  "humanize>=4.12.3,<5.0",
  "jmcomic>=2.6.6",
  "msgspec[toml,yaml]>=0.19.0",
  "paramiko>=4.0.0",
  "mcstatus>=12.0.5",
  "playwright==1.55.0", # pinned for docker image
  "pydantic>=2.11.7",
  "tarina>=0.6.8",
  # plugins
  "nonebot-plugin-exe-code @ file:///${PROJECT_ROOT}/external/nonebot-plugin-exe-code",
  "nonebot-plugin-talk-stats @ file:///${PROJECT_ROOT}/external/nonebot-plugin-talk-stats",
  "nonebot-plugin-werewolf @ file:///${PROJECT_ROOT}/external/nonebot-plugin-werewolf",
  "nonebot-plugin-zssm @ file:///${PROJECT_ROOT}/external/nonebot-plugin-zssm",
]

[dependency-groups]
dev = [
  "ruff>=0.12.10",
  "basedpyright>=1.31.3",
  "nonebot-adapter-console>=0.9.0",
  "nonebot-plugin-orm[aiosqlite]>=0.7.7",
  "nonebug>=0.4.3",
  "pytest>=8.4.1",
  "pytest-mock>=3.14.1",
  "pre-commit>=4.3.0",
  "respx>=0.22.0",
]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
  "F",     # Pyflakes
  "W",     # pycodestyle warnings
  "E",     # pycodestyle errors
  "UP",    # pyupgrade
  "RUF",   # Ruff-specific rules
  "ERA",   # eradicate
  "FLY",   # flynt
  "FURB",  # refurb
  "I",     # isort
  "N",     # PEP8-naming
  "PD",    # pandas-vet
  "PERF",  # Perflint
  "PGH",   # pygrep-hooks
  "TRY",   # tryceratops
  "ASYNC", # flake8-async
  "ANN",   # flake8-annotations
  "S",     # flake8-bandit
  # "FBT",   # flake8-boolean-trap
  "A",     # flake8-builtins
  "B",     # flake8-bugbear
  "COM",   # flake8-commas
  "C4",    # flake8-comprehensions
  "DTZ",   # flake8-datetimez
  "T10",   # flake8-debugger
  "EM",    # flake8-errmsg
  "FA",    # flake8-future-annotations
  "ISC",   # flake8-implicit-str-concat
  "ICN",   # flake8-import-conventions
  "PIE",   # flake8-pie
  "T20",   # flake8-print
  "PYI",   # flake8-pyi
  "PT",    # flake8-pytest-style
  "Q",     # flake8-quotes
  "RSE",   # flake8-raise
  "RET",   # flake8-return
  "SLF",   # flake8-self
  "SIM",   # flake8-simplify
  "SLOT",  # flake8-slots
  "TID",   # flake8-tidy-imports
  "ARG",   # flake8-unused-arguments
  "PTH",   # flake8-use-pathlib
  "TCH",   # flake8-type-checking
  "YTT",   # flake8-2020
]
ignore = [
  "E402",   # module-import-not-at-top-of-file
  "RUF001", # ambiguous-unicode-character-string
  "RUF002", # ambiguous-unicode-character-docstring
  "RUF003", # ambiguous-unicode-character-comment
  "ANN401", # any-type
  "TRY003", # raise-vanilla-args
  "TRY301", # raise-within-try
  "EM102",  # f-string-in-exception
  "COM812", # missing-trailing-comma
  "ERA001", # commented-out-code
  "EM101",  # raw-string-in-exception
  "TID252", # relative-imports
  "ISC001", # single-line-implicit-string-concatenation
  "DTZ005", # call-datetime-now-without-tzinfo
  "S311",   # suspicious-non-cryptographic-random-usage
  "PGH003", # blanket-type-ignore
  "N818",   # error-suffix-on-exception-name  # stupid
  "PYI041", # redundant-numeric-union  # int | float
  "S101",   # assert
  "B911",   # batched-without-explicit-strict
  "PGH004", # blanket-noqa
  "PLC",
]

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id", "dir"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["nonebot_plugin_alconna.Query"]

[tool.ruff.lint.isort]
known-first-party = ["src", "tests/*"]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonVersion = "3.14"
pythonPlatform = "All"
typeCheckingMode = "standard"
ignore = [
  ".venv",
  "data",       # nbp-orm migration scripts
  "**/build",
  "**/scripts",
]
disableBytesTypePromotions = true
